+++ src/api/caidapp/views_tools.py
@@ -1,4 +1,7 @@
-
+from io import BytesIO
+import datetime
+import pandas as pd
+from django.http import HttpResponse